don't validate associations that haven't been loaded.
if the association hasn't loaded, we assume it hasn't changed. this makes
validations (and therefore saves) a bit faster without spiderwebbing out
to encompass large collections. it also reduces the risk of failing to
save a record because an existing associated record is oddly invalid with
no relevance to the current situation.